fix: Home button does not work in the toolbar; refactor OverlayTab
