+++ src/contentscript/overlay/root/components/OverlayTab/styles/OverlayTab.module.scss
@@ -1,4 +1,4 @@
-@import '../../assets/styles/color.variables';
+@import '../../../assets/styles/color.variables';
 @import './adaptive';
 @import './menuWidgets';
