+++ src/contentscript/overlay/root/components/OverlayTab/styles/adaptive.scss
@@ -1,4 +1,4 @@
-@import '../../assets/styles/color.variables';
+@import '../../../assets/styles/color.variables';
 
 @media (min-width: 320px) and (max-width: 640px) {
   .tab {
