+++ src/contentscript/overlay/root/components/OverlayTab/styles/menuWidgets.scss
@@ -1,6 +1,6 @@
-@import '../../assets/styles/null';
-@import '../../assets/styles/fonts';
-@import '../../assets/styles/color.variables';
+@import '../../../assets/styles/null';
+@import '../../../assets/styles/fonts';
+@import '../../../assets/styles/color.variables';
 
 .menuWidgets {
   position: absolute;
